feat: 문제 페이지에서 사용된 UI 컴포넌트화
code, description 폴더 생성
CodeFooter, DescriptionFooter, CodeController 컴포넌트화

Related to: #22
