+++ app/problems/[id]/page.tsx
@@ -289,11 +289,7 @@ export default function Problem({
 
   return (
     <div className="flex h-screen gap-2 mt-3">
-      <ProblemTemplate
-        label="Description"
-        svg="description"
-        height="h-[calc(100%-88px)]"
-      >
+      <ProblemTemplate label="Description" height="h-[calc(100%-88px)]">
         <ProblemDescription
           problemNum={problemNum}
           title={title}
@@ -304,8 +300,9 @@ export default function Problem({
           constraints={constraints}
         />
       </ProblemTemplate>
+
       <div className="flex flex-col gap-2 h-[calc(100%-88px)]">
-        <ProblemTemplate label="Code" svg="solving" height="h-full">
+        <ProblemTemplate label="Code" height="h-full">
           <ProblemSolving id={id} />
         </ProblemTemplate>
         <ProblemTest />
