+++ containers/problems/code/CodeFooter.tsx
@@ -1,0 +1,7 @@
+export default function CodeFooter() {
+  return (
+    <footer className="h-6 pr-4 text-sm text-right text-subtext dark:text-subtextDark">
+      Ln 1, Col 1
+    </footer>
+  );
+}
