+++ containers/problems/code/index.tsx
@@ -1,0 +1,5 @@
+import CodeFooter from './CodeFooter';
+import CodeController from './CodeController';
+import ProblemSolving from './ProblemSolving';
+
+export { CodeFooter, CodeController, ProblemSolving };
