+++ containers/problems/description/index.tsx
@@ -1,0 +1,4 @@
+import DescriptionFooter from './DescriptionFooter';
+import ProblemDescription from './ProblemDescription';
+
+export { DescriptionFooter, ProblemDescription };
