+++ containers/problems/problemTest/index.tsx
@@ -1,0 +1,3 @@
+import ProblemTest from './ProblemTest';
+
+export default ProblemTest;
